+++ tests/unit/models/DefaultConfiguration.spec.js
@@ -6,6 +6,7 @@ describe('Test class: DefaultConfiguration', () => {
       const config = new DefaultConfiguration();
 
       expect(config.editor).toEqual({ syntax: null });
+      expect(config.restrictiveFolder).toBeNull();
       expect(config.defaultFileName).toBeNull();
       expect(config.defaultFileExtension).toBeNull();
       expect(config.tags).toEqual([]);
@@ -15,6 +16,7 @@ describe('Test class: DefaultConfiguration', () => {
       const config = new DefaultConfiguration({});
 
       expect(config.editor).toEqual({ syntax: null });
+      expect(config.restrictiveFolder).toBeNull();
       expect(config.defaultFileName).toBeNull();
       expect(config.defaultFileExtension).toBeNull();
       expect(config.tags).toEqual([]);
@@ -25,12 +27,14 @@ describe('Test class: DefaultConfiguration', () => {
         editor: {
           syntax: true,
         },
+        restrictiveFolder: 'root',
         defaultFileName: 'test',
         defaultFileExtension: '.test',
         tags: ['test'],
       });
 
       expect(config.editor).toEqual({ syntax: true });
+      expect(config.restrictiveFolder).toEqual('root');
       expect(config.defaultFileName).toEqual('test');
       expect(config.defaultFileExtension).toEqual('.test');
       expect(config.tags).toEqual(['test']);